requirements.txt
# Requirements for Master's Thesis Research
# Synergistic Effects of Sampling Methods and AUC Optimization

# Core ML libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0

# LibAUC for AUC optimization
libauc>=1.2.0

# Imbalanced learning
imbalanced-learn>=0.8.0

# Configuration management
PyYAML>=6.0

# Image processing
Pillow>=8.0.0

# Additional utilities
tqdm>=4.62.0